CR and modularize logger
