+++ eslint.config.js
@@ -214,12 +214,10 @@ const importConfig = {
           {
             group: 'external',
             pattern: '@**',
-            position: 'before',
           },
           {
             group: 'internal',
             pattern: '(agent|config|server|utils)/**',
-            position: 'before',
           },
         ],
         pathGroupsExcludedImportTypes: ['react'],
